+++ src/viewport/Manipulator.cpp
@@ -75,13 +75,13 @@ void Manipulator::draw(const SP<Operations> &operations, const Camera &camera) {
         return;
     }
 
-    auto transforms = std::array<mat4, 3> {
+    auto transforms = std::array<dmat4, 3> {
             mat4(1), // x
             mat4(0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1), // y
             mat4(0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1), // z
     };
-    auto colors = std::array<vec3, 3> {
-            vec3(1, 0, 0), vec3(0, 1, 0), vec3(0, 0, 1)
+    auto colors = std::array<dvec3, 3> {
+            dvec3(1, 0, 0), dvec3(0, 1, 0), dvec3(0, 0, 1)
     };
 
     for (size_t i = 0; i < 3; ++i) {
@@ -116,7 +116,7 @@ bool Manipulator::mouseMove(QMouseEvent *event, vec2 pos, const Camera &camera) 
     if (!isInScreen) {
         return false;
     }
-    emit onDrag(vec3(t - _initialDragValue, 0, 0));
+    emit onDrag(dvec3(t - _initialDragValue, 0, 0));
     return true;
 }
 
@@ -129,34 +129,34 @@ bool Manipulator::mouseRelease(QMouseEvent *event, vec2 pos, const Camera &camer
     return true;
 }
 
-std::pair<mat4, bool> Manipulator::manipulatorToWorldMatrix(const Camera &camera) const {
+std::pair<dmat4, bool> Manipulator::manipulatorToWorldMatrix(const Camera &camera) const {
     auto [screenPos, isInScreen] = camera.mapWorldToScreen(_targetPosition);
     if (!isInScreen){
         return {mat4(), false};
     }
-    vec3 screenPosFixedDepth(screenPos.xy, 0.5f);
-    vec3 positionFixedDepth_worldSpace = camera.mapScreenToWorld(screenPosFixedDepth);
+    dvec3 screenPosFixedDepth(screenPos.xy, 0.5);
+    dvec3 positionFixedDepth_worldSpace = camera.mapScreenToWorld(screenPosFixedDepth);
 
-    float scale = 1.f / float(camera.viewSize().y) * 10.f;
+    double scale = 1.0 / double(camera.viewSize().y) * 10.0;
 
-    mat4 modelMatrix = glm::scale(glm::translate(glm::mat4(1), positionFixedDepth_worldSpace), vec3(scale));
+    dmat4 modelMatrix = glm::scale(glm::translate(glm::dmat4(1), positionFixedDepth_worldSpace), dvec3(scale));
     return {modelMatrix, true};
 }
 
-std::tuple<float, float, bool> Manipulator::distanceFromArrow(vec2 screenPos, const Camera& camera) {
+std::tuple<double, double, bool> Manipulator::distanceFromArrow(dvec2 screenPos, const Camera& camera) {
     auto [manipulatorToWorld, isInScreen] = this->manipulatorToWorldMatrix(camera);
     if (!isInScreen) {
         return {0, 0, false};
     }
 
-    vec3 front = camera.mapScreenToCamera(vec3(screenPos, 0));
-    vec3 back = camera.mapScreenToCamera(vec3(screenPos, 1));
+    dvec3 front = camera.mapScreenToCamera(vec3(screenPos, 0));
+    dvec3 back = camera.mapScreenToCamera(vec3(screenPos, 1));
 
-    mat4 manipulatorToCamera = camera.worldToCameraMatrix() * manipulatorToWorld;
+    dmat4 manipulatorToCamera = camera.worldToCameraMatrix() * manipulatorToWorld;
 
-    vec3 arrowXDirection = manipulatorToCamera[0].xyz;
-    float scale = glm::length(arrowXDirection);
-    vec3 arrowCenter = manipulatorToCamera[3].xyz;
+    dvec3 arrowXDirection = manipulatorToCamera[0].xyz;
+    double scale = glm::length(arrowXDirection);
+    dvec3 arrowCenter = manipulatorToCamera[3].xyz;
 
     Line mouseRay(front, back);
     Line arrowRay(arrowCenter, arrowCenter + arrowXDirection);
