+++ src/viewport/operations/DrawSolid.cpp
@@ -10,7 +10,7 @@ DrawSolid::DrawSolid() :
 {
 }
 
-void DrawSolid::draw(const SP<VAO> &vao, const glm::mat4 &matrix, const Camera &camera, glm::vec3 diffuse, glm::vec3 ambient) {
+void DrawSolid::draw(const SP<VAO> &vao, const glm::dmat4 &matrix, const Camera &camera, glm::dvec3 diffuse, glm::dvec3 ambient) {
     _shader.bind();
     _shader.setUniform("diffuse", diffuse);
     _shader.setUniform("ambient", ambient);
