+++ src/viewport/operations/DrawSolid.hpp
@@ -10,8 +10,8 @@ class DrawSolid {
 public:
     DrawSolid();
 
-    void draw(const SP<VAO>& vao, const glm::mat4& matrix, const Camera& camera,
-              glm::vec3 diffuse, glm::vec3 ambient);
+    void draw(const SP<VAO>& vao, const glm::dmat4 &matrix, const Camera& camera,
+              glm::dvec3 diffuse, glm::dvec3 ambient);
 
 private:
     Shader _shader;
